feat: expose common string utilities
