+++ packages/platform-core/__tests__/createShopUtils.test.ts
@@ -1,12 +1,6 @@
 // packages/platform-core/__tests__/createShopUtils.test.ts
 import fs from "fs";
-import {
-  copyTemplate,
-  loadBaseTokens,
-  slugify,
-  genSecret,
-  fillLocales,
-} from "../src/createShop/utils";
+import { copyTemplate, loadBaseTokens } from "../src/createShop/utils";
 
 describe("createShop utils", () => {
   beforeEach(() => {
@@ -39,18 +33,4 @@ describe("createShop utils", () => {
     expect(tokens["--color-bg"]).toBeDefined();
   });
 
-  it("slugifies strings", () => {
-    expect(slugify(" Hello World! ")).toBe("hello-world");
-  });
-
-  it("generates secrets of correct length", () => {
-    const secret = genSecret(8);
-    expect(secret).toHaveLength(16);
-  });
-
-  it("fills locales with fallback", () => {
-    const result = fillLocales({ en: "Hello" }, "Hi");
-    expect(result.en).toBe("Hello");
-    expect(result.de).toBe("Hi");
-  });
 });
