+++ packages/platform-core/src/createShop.ts
@@ -9,13 +9,8 @@ import { join } from "path";
 import { ulid } from "ulid";
 import { z } from "zod";
 import { validateShopName } from "./shops";
-import {
-  copyTemplate,
-  loadBaseTokens,
-  slugify,
-  genSecret,
-  fillLocales,
-} from "./createShop/utils";
+import { copyTemplate, loadBaseTokens } from "./createShop/utils";
+import { slugify, genSecret, fillLocales } from "./utils";
 import { loadThemeTokensNode } from "./themeTokens";
 import { nowIso } from "@shared/date";
 import { defaultFilterMappings } from "./defaultFilterMappings";
