+++ packages/platform-core/src/index.ts
@@ -5,3 +5,4 @@ export * from "./contexts/LayoutContext";
 export * from "./contexts/ThemeContext";
 export * from "./defaultFilterMappings";
 export * from "./themeTokens";
+export * from "./utils";
