+++ packages/platform-core/src/utils/index.ts
@@ -1,0 +1,6 @@
+// packages-platform-core/utils/index.ts
+
+export * from "./getShopFromPath";
+export * from "./replaceShopInPath";
+export * from "./string";
+
